requirements.txt
confusable-homoglyphs==3.2.0
Django==2.2.1
django-debug-toolbar==1.11
django-registration==3.0.1
djangorestframework==3.9.4
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
